# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "\u623f\u5c4b\u79df\u8d41AI\u8425\u9500\u7cfb\u7edf",
        "description": "AI\u9a71\u52a8\u7684\u623f\u6e90\u6587\u6848\u751f\u6210\u4e0e\u591a\u5e73\u53f0\u53d1\u5e03\u7cfb\u7edf",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/publish/xiaohongshu": {
            "post": {
                "tags": [
                    "\u53d1\u5e03\u7ba1\u7406"
                ],
                "summary": "Publish To Xiaohongshu",
                "description": "\u53d1\u5e03\u6587\u6848\u5230\u5c0f\u7ea2\u4e66\n\n- **script_id**: \u6587\u6848ID\n- **images**: \u56fe\u7247\u8def\u5f84\u5217\u8868\n\n\u8fd4\u56de\uff1a\u53d1\u5e03\u7ed3\u679c",
                "operationId": "publish_to_xiaohongshu_api_v1_publish_xiaohongshu_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublishResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/publish/wechat": {
            "post": {
                "tags": [
                    "\u53d1\u5e03\u7ba1\u7406"
                ],
                "summary": "Create Wechat Draft",
                "description": "\u521b\u5efa\u516c\u4f17\u53f7\u8349\u7a3f\n\n- **script_id**: \u6587\u6848ID\n- **images**: \u56fe\u7247\u8def\u5f84\u5217\u8868\n\n\u8fd4\u56de\uff1a\u521b\u5efa\u7ed3\u679c",
                "operationId": "create_wechat_draft_api_v1_publish_wechat_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublishResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/publish/logs": {
            "get": {
                "tags": [
                    "\u53d1\u5e03\u7ba1\u7406"
                ],
                "summary": "Get Publish Logs",
                "description": "\u83b7\u53d6\u53d1\u5e03\u8bb0\u5f55\n\n- **house_id**: \u7b5b\u9009\u6307\u5b9a\u623f\u6e90\u7684\u8bb0\u5f55\uff08\u53ef\u9009\uff09\n\n\u8fd4\u56de\uff1a\u53d1\u5e03\u8bb0\u5f55\u5217\u8868",
                "operationId": "get_publish_logs_api_v1_publish_logs_get",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "title": "House Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PublishLogResponse"
                                    },
                                    "title": "Response Get Publish Logs Api V1 Publish Logs Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PublishLogResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "house_id": {
                        "type": "integer",
                        "title": "House Id"
                    },
                    "script_id": {
                        "type": "integer",
                        "title": "Script Id"
                    },
                    "platform": {
                        "type": "string",
                        "title": "Platform"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "error_msg": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Msg"
                    },
                    "xhs_note_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Xhs Note Id"
                    },
                    "wechat_media_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Wechat Media Id"
                    },
                    "published_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Published At"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "house_id",
                    "script_id",
                    "platform",
                    "status",
                    "created_at"
                ],
                "title": "PublishLogResponse",
                "description": "\u53d1\u5e03\u8bb0\u5f55\u54cd\u5e94Schema"
            },
            "PublishRequest": {
                "properties": {
                    "script_id": {
                        "type": "integer",
                        "title": "Script Id",
                        "description": "\u6587\u6848ID"
                    },
                    "images": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Images",
                        "description": "\u56fe\u7247\u8def\u5f84\u5217\u8868"
                    }
                },
                "type": "object",
                "required": [
                    "script_id",
                    "images"
                ],
                "title": "PublishRequest",
                "description": "\u53d1\u5e03\u8bf7\u6c42Schema\n\u7528\u4e8e\uff1a\u53d1\u5e03\u6587\u6848\u5230\u5404\u5e73\u53f0"
            },
            "PublishResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "platform": {
                        "type": "string",
                        "title": "Platform"
                    },
                    "note_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Note Id"
                    },
                    "media_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Media Id"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    }
                },
                "type": "object",
                "required": [
                    "success",
                    "platform"
                ],
                "title": "PublishResponse",
                "description": "\u53d1\u5e03\u54cd\u5e94Schema\n\u7528\u4e8e\uff1a\u8fd4\u56de\u53d1\u5e03\u7ed3\u679c"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}